+++ tests/testthat/test-#15.R
@@ -7,15 +7,15 @@ testthat::test_that("simsum with 'true' as a column returns the same results as 
 
   s1 <- simsum(data = tt, estvarname = "diff", true = "true", se = "se", ci.limits = c("lower", "upper"), methodvar = "method", by = "dgm")
   s2 <- simsum(data = tt, estvarname = "diff", true = -1, se = "se", ci.limits = c("lower", "upper"), methodvar = "method", by = "dgm")
-  testthat::expect_identical(object = get_data(s1), expected = get_data(s2))
+  testthat::expect_identical(object = tidy(s1), expected = tidy(s2))
 
   s1 <- simsum(data = tt, estvarname = "diff", true = "true", se = "se", methodvar = "method", by = "dgm")
   s2 <- simsum(data = tt, estvarname = "diff", true = -1, se = "se", methodvar = "method", by = "dgm")
-  testthat::expect_identical(object = get_data(s1), expected = get_data(s2))
+  testthat::expect_identical(object = tidy(s1), expected = tidy(s2))
 
   s1 <- simsum(data = tt, estvarname = "diff", true = "true", se = "se", ci.limits = c(-1.5, -0.5), methodvar = "method", by = "dgm")
   s2 <- simsum(data = tt, estvarname = "diff", true = -1, se = "se", ci.limits = c(-1.5, -0.5), methodvar = "method", by = "dgm")
-  testthat::expect_identical(object = get_data(s1), expected = get_data(s2))
+  testthat::expect_identical(object = tidy(s1), expected = tidy(s2))
 })
 
 testthat::test_that("simsum with 'true' as a column returns different results compared to using a scalar value if each row has a different value", {
@@ -24,15 +24,15 @@ testthat::test_that("simsum with 'true' as a column returns different results co
 
   s1 <- simsum(data = tt, estvarname = "diff", true = "true", se = "se", ci.limits = c("lower", "upper"), methodvar = "method", by = "dgm")
   s2 <- simsum(data = tt, estvarname = "diff", true = -1, se = "se", ci.limits = c("lower", "upper"), methodvar = "method", by = "dgm")
-  testthat::expect_false(object = identical(get_data(s1), get_data(s2)))
+  testthat::expect_false(object = identical(tidy(s1), tidy(s2)))
 
   s1 <- simsum(data = tt, estvarname = "diff", true = "true", se = "se", methodvar = "method", by = "dgm")
   s2 <- simsum(data = tt, estvarname = "diff", true = -1, se = "se", methodvar = "method", by = "dgm")
-  testthat::expect_false(object = identical(get_data(s1), get_data(s2)))
+  testthat::expect_false(object = identical(tidy(s1), tidy(s2)))
 
   s1 <- simsum(data = tt, estvarname = "diff", true = "true", se = "se", ci.limits = c(-1.5, -0.5), methodvar = "method", by = "dgm")
   s2 <- simsum(data = tt, estvarname = "diff", true = -1, se = "se", ci.limits = c(-1.5, -0.5), methodvar = "method", by = "dgm")
-  testthat::expect_false(object = identical(get_data(s1), get_data(s2)))
+  testthat::expect_false(object = identical(tidy(s1), tidy(s2)))
 })
 
 testthat::test_that("multisimsum with 'true' as a column returns the same results as using a scalar value if each row has the same value", {
@@ -45,7 +45,7 @@ testthat::test_that("multisimsum with 'true' as a column returns the same result
 
   m1 <- multisimsum(data = tt, estvarname = "diff", par = "par", true = c(diff1 = -1, diff2 = 0), se = "se", ci.limits = c("lower", "upper"), methodvar = "method", by = "dgm")
   m2 <- multisimsum(data = tt, estvarname = "diff", par = "par", true = "true", se = "se", ci.limits = c("lower", "upper"), methodvar = "method", by = "dgm")
-  testthat::expect_identical(object = get_data(m1), expected = get_data(m2))
+  testthat::expect_identical(object = tidy(m1), expected = tidy(m2))
 })
 
 testthat::test_that("multisimsum with 'true' as a column returns different results compared to using a scalar value if each row has a different value", {
@@ -58,5 +58,5 @@ testthat::test_that("multisimsum with 'true' as a column returns different resul
 
   m1 <- multisimsum(data = tt, estvarname = "diff", par = "par", true = c(diff1 = -1, diff2 = 0), se = "se", ci.limits = c("lower", "upper"), methodvar = "method", by = "dgm")
   m2 <- multisimsum(data = tt, estvarname = "diff", par = "par", true = "true", se = "se", ci.limits = c("lower", "upper"), methodvar = "method", by = "dgm")
-  testthat::expect_false(object = identical(get_data(m1), get_data(m2)))
+  testthat::expect_false(object = identical(tidy(m1), tidy(m2)))
 })